Improve Whizbang auto deck selection
